requirements.txt
# jenkins_monitor/requirements.txt

Flask
python-jenkins